Turret Firing Bug Fix
Turret targeting was getting the turret mesh as the target. That is why it wasn't firing. It was intended to set via BP by making Visibility channel to Overlap.

But I coded it, so, it automatically set it at the begining. Tested it with a new BP called BP_TurretZ

Closes #9